+++ client/src/components/groupserver/GroupServer.jsx
@@ -1,18 +1,14 @@
-import React, { useState, useEffect, useContext } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Redirect, useParams } from 'react-router-dom';
-import { Context } from '../../contexts/Store';
-import { UserContext } from '../../contexts/user-context';
 import TextChannel from './textchat/TextChannel';
 import ServerSidebar from '../ServerSidebar';
 import Loading from '../Loading';
 
 export default function GroupServer() {
-  const [state, setState] = useContext(Context);
-  const [user, setUser] = useContext(UserContext);
   const [loading, setLoading] = useState(true);
   const [userAccess, setUserAccess] = useState(false);
   const params = useParams();
-  const { groupServerId, textChannelId } = useParams();
+  const { groupServerId } = useParams();
 
   // Upon intialization, verify that the user is authorized to be in the group server
   // If the user is not authorized, userAccess = false which redirects the user to "/dashboard"
@@ -24,12 +20,6 @@ export default function GroupServer() {
           headers: {
             'Content-Type': 'application/json',
           },
-          /*
-          body: JSON.stringify({
-            groupServerId,
-            textChannelId: textChannelId || Object.keys(state.groupServers[groupServerId].textChannels)[0],
-          }),
-          */
         }).then((response) => {
           setUserAccess(response.status === 200);
         });
